% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{currentVersion<-}
\alias{currentVersion<-}
\title{Set current version}
\usage{
currentVersion(this) <- value
}
\arguments{
\item{this}{Object of class ''}

\item{value}{A \code{numeric} value}
}
\value{
Object of class ''
}
\description{
Set the highest version that is supported by the package.
}
\details{
The value for the highest version is a static variable. It is
initialized in the initialization method when an instance of a
\code{MiRNANameConverter} class is created.
}
\author{
Stefan Haunsberger
}
