% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{currentVersion}
\alias{currentVersion}
\alias{currentVersion,MiRNANameConverter-method}
\title{Get current version}
\usage{
currentVersion(this)

\S4method{currentVersion}{MiRNANameConverter}(this)
}
\arguments{
\item{this}{Object of class \code{MiRNAmeConverter}}
}
\value{
A numeric value
}
\description{
This function returns the highest miRBase version
that is provided by the package.
}
\details{
The maximum miRBase version of the package is evaluated and set
in the object initialization.
}
\section{Methods (by class)}{
\itemize{
\item \code{MiRNANameConverter}: Retrieve highest supported miRBase version
}}
\examples{
nc = MiRNANameConverter(); # Instance of class 'MiRNANameConverter'
currentVersion(nc);
}
\author{
Stefan Haunsberger
}

