% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.miRNAs}
\alias{example.miRNAs}
\title{miRNA names.}
\format{A \code{character} vector containing names.}
\usage{
example.miRNAs
}
\description{
Sample names including miRNA names, non-miRNA names and other.
It also includes duplicates.
}
\keyword{datasets}

