% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiRNANameConverter.R
\docType{methods}
\name{MiRNANameConverter,ANY-method}
\alias{MiRNANameConverter,ANY-method}
\title{MiRNANameConverter constructor}
\usage{
\S4method{MiRNANameConverter}{ANY}()
}
\value{
an object of class 'MiRNANameConverter'
}
\description{
This function returns an instance of a MiRNANAmeConverter class.
}
\details{
This function initializes an object of the class MiRNANameConverter. It is a
wrapper for \code{new()}.
}
\examples{
nc = MiRNANameConverter() # Instance of class 'MiRNANameConverter'

}
\author{
Stefan Haunsberger
}
\seealso{
\code{\link{new}}
}

