% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiRNANameConverter.R
\name{MiRNANameConverter}
\alias{MiRNANameConverter}
\title{Instantiate from MiRNANameConverter class}
\usage{
MiRNANameConverter(...)
}
\arguments{
\item{...}{any optional arguments}
}
\description{
This function returns back an instance of a
\emph{MiRNANAmeConverter} object.
}
\section{Slots}{

\describe{
\item{\code{.dbconn}}{Database connection}

\item{\code{.currentVersion}}{Current miRBase version}

\item{\code{.validVersions}}{Valid/Supported miRBase versions}

\item{\code{.nOrganisms}}{Number of different organisms supported}

\item{\code{.nTotalEntries}}{Total number of mature miRNA names among all provided
miRBase release versions in the \emph{miRBaseVersions.db} package.}

\item{\code{.validOrganisms}}{Valid organisms}
}}
\author{
Stefan Haunsberger
}

