% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{nOrganisms<-}
\alias{nOrganisms<-}
\title{Set number of organisms}
\usage{
nOrganisms(this) <- value
}
\arguments{
\item{this}{Object of class \code{MiRNANameConverter}}

\item{value}{An \code{integer} value}
}
\value{
A \code{MiRNANameConverter} object
}
\description{
This function sets the number of different organisms
that are provided by the package.
}
\details{
The number of different organisms is evaluated and set
in the object initialization.
}
\author{
Stefan Haunsberger
}

