% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{nTotalEntries<-}
\alias{nTotalEntries<-}
\title{Set total number database entries}
\usage{
nTotalEntries(this) <- value
}
\arguments{
\item{this}{Object of class \code{MiRNANameConverter}}

\item{value}{An \code{integer} value}
}
\value{
A \code{MiRNANameConverter} object
}
\description{
This function sets the total number of entries contained
in the \code{mimat} table. The number is the sum of the entries of all
miRBase versions provided by the package.
}
\details{
The total number is evaluated and set in the object initialization.
}
\author{
Stefan Haunsberger
}
