% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{nTotalEntries}
\alias{nTotalEntries}
\alias{nTotalEntries,MiRNANameConverter-method}
\title{Get total number database entries}
\usage{
nTotalEntries(this)

\S4method{nTotalEntries}{MiRNANameConverter}(this)
}
\arguments{
\item{this}{Object of class \code{MiRNAmeConverter}}
}
\value{
A numeric value
}
\description{
This function returns the total number of entries contained
in the \code{mimat} table. The number is the sum of the entries of all
miRBase versions provided by the package.
}
\details{
The total number is evaluated and set in the object initialization.
}
\section{Methods (by class)}{
\itemize{
\item \code{MiRNANameConverter}: Retrieve total number of miRNA entries
}}

\examples{
nc = MiRNANameConverter(); # Instance of class 'MiRNANameConverter'
nTotalEntries(nc);
}
\author{
Stefan Haunsberger
}
