% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiRNANameConverter.R
\docType{methods}
\name{show,MiRNANameConverter-method}
\alias{show,MiRNANameConverter-method}
\title{Show-method}
\usage{
\S4method{show}{MiRNANameConverter}(object)
}
\arguments{
\item{object}{Object of class \code{MiRNANameConverter}}
}
\description{
This function prints object specific information
Show-method

This function prints object specific information
}
\details{
This function prints some information to the console.
}
\seealso{
\code{\link{show}}
}
\author{
Stefan Haunsberger
}
