% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{validOrganisms<-}
\alias{validOrganisms<-}
\title{Set valid organisms}
\usage{
validOrganisms(this) <- value
}
\arguments{
\item{this}{Object of class \code{MiRNANameConverter}}

\item{value}{A \code{character} vector}
}
\value{
A \code{MiRNANameConverter} object
}
\description{
This function sets all organisms where mature miRNA names
are available in any of the provided miRBase versions.
}
\details{
The valid organisms are evaluated and set in the object initialization.
}
\author{
Stefan Haunsberger
}

