% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{validOrganisms}
\alias{validOrganisms}
\alias{validOrganisms,MiRNANameConverter-method}
\title{Get valid organisms}
\usage{
validOrganisms(this)

\S4method{validOrganisms}{MiRNANameConverter}(this)
}
\arguments{
\item{this}{Object of class \code{MiRNAmeConverter}}
}
\value{
A numeric value
}
\description{
This function returns all organisms where mature miRNA names
are available in any of the provided miRBase versions.
}
\details{
The valid organisms are evaluated and set in the object initialization.
}
\section{Methods (by class)}{
\itemize{
\item \code{MiRNANameConverter}: Retrieve all supported organisms
}}

\examples{
nc = MiRNANameConverter(); # Instance of class 'MiRNANameConverter'
nOrganisms(nc);
}
\author{
Stefan Haunsberger
}
