% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters.R
\name{validVersions<-}
\alias{validVersions<-}
\title{Set valid versions}
\usage{
validVersions(this) <- value
}
\arguments{
\item{this}{Object of class \code{MiRNANameConverter}}

\item{value}{A vector of numeric values}
}
\value{
A \code{MiRNANameConverter} object
}
\description{
Set version values that are supported by the package.
}
\details{
The value for the highest versions is a static variable. It is
initialized in the initialization method when an instance of a
\code{MiRNANameConverter} class is created.
}
\author{
Stefan Haunsberger
}

