% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\docType{methods}
\name{validVersions}
\alias{validVersions}
\alias{validVersions,MiRNANameConverter-method}
\title{Get valid versions}
\usage{
validVersions(this)

\S4method{validVersions}{MiRNANameConverter}(this)
}
\arguments{
\item{this}{Object of class \code{MiRNAmeConverter}}
}
\value{
A numeric vector
}
\description{
This function returns all valid miRBase versions
provided by the package.
}
\section{Methods (by class)}{
\itemize{
\item \code{MiRNANameConverter}: Retrieve supported miRBase versions
}}
\examples{
validVersions
nc = MiRNANameConverter(); # Instance of class 'MiRNANameConverter'
validVersions(nc);
}
\author{
Stefan Haunsberger
}

